refactor: HomeView UI 컴포넌트 등 파일 분리
